Changes in the hyperlinks in index.ipynb
After downloading the repository and open it with jupyter notebook, clicking in the hyperlinks showed error 404, I think it's because they referred to the page in GitHub directly and not to a directory relative to index.ipynb. I changed all the links to refer relatively to index.ipynb instead of Github directly and now it's working for me.